Agreement: Rename `entities` dir to `models`
