+++ packages/connect-agreement/src/index.ts
@@ -1,6 +1,6 @@
-export { default as Agreement } from './entities/Agreement'
-export { default as Signer } from './entities/Signer'
-export { default as Signature } from './entities/Signature'
-export { default as Version } from './entities/Version'
+export { default as Agreement } from './models/Agreement'
+export { default as Signer } from './models/Signer'
+export { default as Signature } from './models/Signature'
+export { default as Version } from './models/Version'
 export { default as AgreementConnectorTheGraph } from './thegraph/connector'
 export { default } from './connect'
