+++ packages/connect-agreement/src/models/Agreement.ts
