+++ packages/connect-agreement/src/models/Signature.ts
