+++ packages/connect-agreement/src/types.ts
@@ -1,8 +1,8 @@
 import { SubscriptionHandler } from '@aragon/connect-types'
 
-import Signer from './entities/Signer'
-import Signature from './entities/Signature'
-import Version from './entities/Version'
+import Signer from './models/Signer'
+import Signature from './models/Signature'
+import Version from './models/Version'
 
 export interface AgreementData {
   id: string
